+++ src/main/java/pl/bartlomiejstepien/mcsm/domain/model/ServerPack.java
@@ -2,16 +2,16 @@ package pl.bartlomiejstepien.mcsm.domain.model;
 
 public class ServerPack
 {
-    private final int id;
+    private final Long id;
     private final String name;
 
-    public ServerPack(final int id, final String name)
+    public ServerPack(final Long id, final String name)
     {
         this.id = id;
         this.name = name;
     }
 
-    public int getId()
+    public Long getId()
     {
         return id;
     }
